refactor(cli): split MCP and CLI into separate entry files; rename subcommands
- src/index.ts is now the MCP-only entry (flag-only, no Commander subcommands). All serve/HTTP transport plumbing lives here.
- src/cli.ts is now the CLI-only entry with three subcommands: sections, search, items. No MCP imports.
- src/commands.ts holds the action handlers (runSections, runSearch, runItems) so tests can import them without triggering arg parsing.
- bin field: context-awesome → build/cli.js, context-awesome-mcp → build/index.js.
- CLI renames: lookup → sections, docs → items. serve removed from CLI entirely.
- Legacy MCP configs that pass serve as an arg still work via .allowUnknownOption().
- README updated with new CLI names and clarified CLI vs MCP split.
- Dockerfile CMD dropped the serve arg.
- cli-search.test.ts import path updated to ../commands.js.
